+++ game/board.py
@@ -70,6 +70,6 @@ class Board:
         elif self.grid[x-1][y-1].tile.letter == tile.letter:
             self.grid[x-1][y-1].noneTile()
             self.grid[x-1][y-1].addTile(tile)
-    
+            
     def getCellInBoard(self, x, y):
         return self.grid[x-1][y-1]